Модифицировал комманду удаления бэкапа

+++ scripts/TouchinBuild/commands/BaseBackupCommand/DeleteBackupCommand.py
@@ -1,3 +1,4 @@
+import os
 import shutil
 from commands.BaseBackupCommand.BaseBackupCommand import BaseBackupCommand
 
@@ -7,4 +8,7 @@ class DeleteBackupCommand(BaseBackupCommand):
 		BaseBackupCommand.__init__(self)
 
 	def execute(self):
+		if not os.path.exists(self.backupDirAbsPath):
+			raise Exception('backup folder: {0} not exists'.format(self.backupDirAbsPath))
+
 		shutil.rmtree(self.backupDirAbsPath, ignore_errors=True)
